requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Migrate==4.1.0
PyMySQL==1.1.1
cryptography==44.0.0
Werkzeug==3.1.3
WTForms==3.2.1
email-validator==2.2.0
Pillow==11.1.0
openpyxl==3.1.5
